requirements.txt
# Runtime
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic-settings==2.13.1
structlog==25.5.0

# Development / testing
pytest==8.3.5
pytest-cov==7.1.0
httpx==0.28.1
mypy==1.20.1
ruff==0.15.11
bandit==1.9.4
pre-commit==4.2.0
